+++ packages/canonical-path/src/index.js
@@ -1,9 +1,13 @@
-const content = require('./content');
-const section = require('./section');
+const platformContent = require('./platform/content');
+const websiteSection = require('./website/section');
 const requestParser = require('./request-parser');
 
 module.exports = {
-  content,
-  section,
+  platformContent,
+  websiteSection,
   requestParser,
+
+  // @todo update refs to these to remove them: BC break, minor this package
+  content: platformContent,
+  section: websiteSection,
 };
