+++ packages/canonical-path/src/platform/content.js
@@ -1,7 +1,20 @@
 const { isFunction: isFn, cleanPath } = require('@base-cms/utils');
 const { get } = require('@base-cms/object-path');
+const { BaseDB } = require('@base-cms/db');
+const { dasherize } = require('@base-cms/inflector');
 
-const pathResolvers = require('./path-resolvers');
+const pathResolvers = {
+  id: content => content._id,
+  slug: content => get(content, 'mutations.Website.slug'),
+  type: content => dasherize(content.type),
+  sectionAlias: async (content, { load }) => {
+    const ref = BaseDB.get(content, 'mutations.Website.primarySection');
+    const id = BaseDB.extractRefId(ref);
+    if (!id) return 'home';
+    const section = await load('websiteSection', id, { alias: 1 }, { status: 1 });
+    return section ? section.alias : 'home';
+  },
+};
 
 const dynamicPageResolvers = {
   alias: content => get(content, 'mutations.Website.alias'),
